Put together all chart-related helper functions
