+++ src/visits/helpers/DefaultChart.tsx
@@ -3,11 +3,11 @@ import { Doughnut, HorizontalBar } from 'react-chartjs-2';
 import { keys, values } from 'ramda';
 import classNames from 'classnames';
 import Chart, { ChartData, ChartDataSets, ChartOptions } from 'chart.js';
-import { fillTheGaps, renderDoughnutChartLabel, renderNonDoughnutChartLabel } from '../../utils/helpers/visits';
+import { fillTheGaps} from '../../utils/helpers/visits';
 import { Stats } from '../types';
 import { prettify } from '../../utils/helpers/numbers';
+import { pointerOnHover, renderDoughnutChartLabel, renderNonDoughnutChartLabel } from '../../utils/helpers/charts';
 import './DefaultChart.scss';
-import { pointerOnHover } from '../../utils/helpers/charts';
 
 export interface DefaultChartProps {
   title: Function | string;
